requirements.txt
fastapi
strawberry-graphql[fastapi]
openai
uvicorn
dotenv